+++ src/sme_contrib/pyvista_utils.py
@@ -1,10 +1,8 @@
 import pyvista as pv
 import numpy as np
-from typing import Callable, Any
 from itertools import cycle
 import matplotlib.colors as mcolors
 import matplotlib.pyplot as plt
-import sme
 
 
 def rgb_to_scalar(img: np.ndarray) -> np.ndarray:
@@ -88,6 +86,7 @@ def find_layout(num_plots: int, portrait: bool = False) -> tuple[int, int]:
         a_1 = int(a - 1)
         b_1 = int(b + 1)
 
+        # make a couple of guesses that are close to the root and select the best one
         guesses = [
             (x, y)
             for x, y in [
@@ -108,171 +107,3 @@ def find_layout(num_plots: int, portrait: bool = False) -> tuple[int, int]:
             if not portrait
             else (np.max(best_guess), np.min(best_guess))
         )
-
-
-def facet_grid(
-    data: dict[str, np.ndarray],
-    plotfuncs: dict[str, Callable],
-    show_cmap: bool = False,
-    cmap: str | np.ndarray | pv.LookupTable = "viridis",
-    portrait: bool = False,
-    linked_views: bool = True,
-    plotter_kwargs: dict = {},
-    plotfuncs_kwargs: dict[str, dict[str, Any]] = {},
-) -> pv.Plotter:
-    """
-    Create a 3D facet plot using PyVista. This follows the seaborn.FacetGrid concept. This function creates a grid of subplots where each subplot is filled by a function in the plotfuncs argument. The keys for plotfuncs and data must be the same, such that plotfuncs can be unambiguously mapped over the data dictionary.
-    Do not attempt to plot 2D images and 3D images into the same facet grid, as this will create odd artifacts and
-    may not work as expected.
-    Parameters:
-    -----------
-    data : dict[str, np.ndarray]
-        A dictionary where keys are labels and values are numpy arrays containing the data to be plotted.
-    plotfuncs : dict[str, Callable]
-        A dictionary where keys are labels and values are functions with signature f(
-                label:str,
-                data:np.ndarray | pyvista.ImageData | pyvista.UniformGrid,
-                plotter:pv.Plotter,
-                panel:tuple[int, int],
-                show_cmap:bool=show_cmap,
-                cmap=cmap,
-                **plotfuncs_kwargs
-            ) -> None
-    show_cmap : bool, optional
-        Whether to show the color map. Default is False.
-    cmap : str | np.ndarray | pv.LookupTable, optional
-        The color map to use. Default is "viridis".
-    portrait : bool, optional
-        Whether to use a portrait layout. Default is False.
-    linked_views : bool, optional
-        Whether to link the views of the subplots. Default is True.
-    plotter_kwargs : dict, optional
-        Additional keyword arguments to pass to the PyVista Plotter.
-    plotfuncs_kwargs : dict[str, dict[str, Any]], optional
-        Additional keyword arguments to pass to each plotting function.
-
-    Returns:
-    --------
-    pv.Plotter
-        The PyVista Plotter object with the created facet plot.
-    """
-    if data.keys() != plotfuncs.keys():
-        raise ValueError(
-            "The keys for the data and plotfuncs dictionaries must be the same."
-        )
-
-    layout = find_layout(len(data), portrait=portrait)
-
-    plotter = pv.Plotter(shape=layout, **plotter_kwargs)
-
-    label = iter(plotfuncs.keys())
-
-    for i in range(layout[0]):
-        for j in range(layout[1]):
-            current_label = next(label)
-            plotfuncs[current_label](
-                current_label,
-                data[current_label],
-                plotter,
-                panel=(i, j),
-                show_cmap=show_cmap,
-                cmap=cmap,
-                **plotfuncs_kwargs.get(current_label, {}),
-            )
-
-    if linked_views:
-        plotter.link_views()
-
-    return plotter
-
-
-def facet_grid_animate(
-    filename: str,
-    data: list[dict[str, np.ndarray]],
-    plotfuncs: dict[str, Callable],
-    show_cmap: bool = False,
-    cmap: str | np.ndarray | pv.LookupTable = "viridis",
-    portrait: bool = False,
-    linked_views: bool = True,
-    titles: list[dict[str, str]] = [],
-    plotter_kwargs: dict = {},
-    plotfuncs_kwargs: dict[str, dict[str, Any]] = {},
-) -> str:
-    """
-    Create a 3D animation from a series of data snapshots using PyVista.
-    This series must be a list of dictionaries with the data for each frame keyed by a label used to title the panel it will be plotted into. The final plot will have as many subplots as there are labels in the data dictionaries. The keys for plotfuncs and data must be the same.
-    Parameters:
-    -----------
-    filename : str
-        The name of the output movie file.
-    data : list[dict[str, np.ndarray]]
-        A list of dictionaries containing the data for each timestep.
-    plotfuncs : dict[str, Callable]
-        A dictionary of plotting functions keyed by data label. The keys for plotfuncs and data must be the same.
-    show_cmap : bool, optional
-        Whether to show the color map (default is False).
-    cmap : str | np.ndarray | pv.LookupTable, optional
-        The colormap to use (default is "viridis").
-    portrait : bool, optional
-        Whether to use portrait layout (default is False).
-    linked_views : bool, optional
-        Whether to link the views of the subplots (default is True).
-    titles : list[dict[str, str]], optional
-        A list of dictionaries containing titles for each subplot (default is an empty list).
-    plotter_kwargs : dict, optional
-        Additional keyword arguments to pass to the PyVista Plotter (default is an empty dictionary).
-    plotfuncs_kwargs : dict[str, dict[str, Any]], optional
-        Additional keyword arguments to pass to each plotting function (default is an empty dictionary).
-    Returns:
-    --------
-    str
-        The filename of the created movie.
-    """
-
-    if len(titles) > 0 and len(titles) != len(data):
-        raise ValueError(
-            "The number of titles must be the same as the number of data dictionaries."
-        )
-
-    if data[0].keys() != plotfuncs.keys():
-        raise ValueError(
-            "The keys for the data and plotfuncs dictionaries must be the same."
-        )
-
-    def create_frame(
-        data_dict: dict[str, np.ndarray], title: dict[str:str], layout=(1, 1)
-    ):
-        label = iter(data_dict.keys())
-        for i in range(layout[0]):
-            for j in range(layout[1]):
-                current_label = next(label)
-                plotfuncs[current_label](
-                    title.get(current_label, current_label),
-                    data_dict[current_label],
-                    plotter,
-                    panel=(i, j),
-                    show_cmap=show_cmap,
-                    cmap=cmap,
-                    **plotfuncs_kwargs.get(current_label, {}),
-                )
-
-        plotter.write_frame()
-
-    layout = find_layout(len(plotfuncs), portrait=portrait)
-
-    plotter = pv.Plotter(shape=layout, **plotter_kwargs)
-
-    plotter.open_movie(filename)
-    create_frame(data[0], titles[0] if len(titles) > 0 else {}, layout)
-
-    if linked_views:
-        plotter.link_views()
-
-    for i, single_timestep_data in enumerate(data[1::]):
-        create_frame(
-            single_timestep_data, titles[i] if len(titles) > 0 else {}, layout=layout
-        )
-
-    plotter.close()
-
-    return filename
