+++ tests/test_pyvista_utils.py
@@ -2,7 +2,6 @@ import sme_contrib.pyvista_utils as pvu
 import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib.colors as mcolors
-import pytest
 
 
 def test_rgb_to_scalar():
@@ -62,130 +61,3 @@ def test_find_layout():
     assert pvu.find_layout(29) == (5, 6)
     assert pvu.find_layout(31) == (5, 7)
     assert pvu.find_layout(31, portrait=True) == (7, 5)
-
-
-def test_facet_grid_3D(exampledata):
-    def plot_bloodvessel(label, data, plotter, panel, **kwargs):
-        plotter.subplot(*panel)
-        plotter.add_mesh(data)
-
-    def plot_brain(label, data, plotter, panel, **kwargs):
-        plotter.subplot(*panel)
-        plotter.add_volume(
-            data,
-            cmap="viridis",
-            opacity="sigmoid",  # Common opacity mapping for volume rendering
-            shade=True,
-            ambient=0.3,
-            diffuse=0.6,
-            specular=0.5,
-        )
-
-    def plot_armadillo(label, data, plotter, panel, **kwargs):
-        plotter.subplot(*panel)
-        plotter.add_mesh(data)
-
-    facetgrid = pvu.facet_grid(
-        data={
-            "armadillo": exampledata["armadillo"],
-            "bloodvessel": exampledata["bloodvessel"],
-            "brain": exampledata["brain"],
-        },
-        plotfuncs={
-            "armadillo": plot_armadillo,
-            "bloodvessel": plot_bloodvessel,
-            "brain": plot_brain,
-        },
-    )
-
-    assert facetgrid.shape == (1, 3)
-
-    with pytest.raises(ValueError):
-        pvu.facet_grid(
-            data={
-                "armadillo": exampledata["armadillo"],
-                "bloodvessel": exampledata["bloodvessel"],
-                "brain": exampledata["brain"],
-            },
-            plotfuncs={
-                "armadillo": plot_armadillo,
-                "bloodvessel": plot_bloodvessel,
-                "wrong_key": plot_brain,
-            },
-        )
-
-
-def test_facet_grid_animation(tmp_path, exampledata):
-    def plot_bloodvessel(label, data, plotter, panel, **kwargs):
-        plotter.subplot(*panel)
-        plotter.add_mesh(data)
-
-    def plot_brain(label, data, plotter, panel, **kwargs):
-        plotter.subplot(*panel)
-        plotter.add_volume(
-            data,
-            cmap="viridis",
-            opacity="sigmoid",  # Common opacity mapping for volume rendering
-            shade=True,
-            ambient=0.3,
-            diffuse=0.6,
-            specular=0.5,
-        )
-
-    def plot_armadillo(label, data, plotter, panel, **kwargs):
-        plotter.subplot(*panel)
-        plotter.add_mesh(data)
-
-    data_for_frames = [
-        {
-            "armadillo": exampledata["armadillo"],
-            "bloodvessel": exampledata["bloodvessel"],
-            "brain": exampledata["brain"],
-        },
-        {
-            "armadillo": exampledata["armadillo"],
-            "bloodvessel": exampledata["bloodvessel"],
-            "brain": exampledata["brain"],
-        },
-        {
-            "armadillo": exampledata["armadillo"],
-            "bloodvessel": exampledata["bloodvessel"],
-            "brain": exampledata["brain"],
-        },
-    ]
-
-    testanimation = pvu.facet_grid_animate(
-        tmp_path / "test.mp4",
-        data=data_for_frames,
-        plotfuncs={
-            "armadillo": plot_armadillo,
-            "bloodvessel": plot_bloodvessel,
-            "brain": plot_brain,
-        },
-    )
-
-    assert testanimation == tmp_path / "test.mp4"
-    assert testanimation.exists()
-
-    with pytest.raises(ValueError):
-        pvu.facet_grid_animate(
-            tmp_path / "test.mp4",
-            data=data_for_frames,
-            plotfuncs={
-                "armadillo": plot_armadillo,
-                "bloodvessel": plot_bloodvessel,
-                "wrong_key": plot_brain,
-            },
-        )
-
-    with pytest.raises(ValueError):
-        pvu.facet_grid_animate(
-            tmp_path / "test.mp4",
-            data=data_for_frames,
-            plotfuncs={
-                "armadillo": plot_armadillo,
-                "bloodvessel": plot_bloodvessel,
-                "brain": plot_brain,
-            },
-            titles=["title1", "title2"],
-        )
